+++ semana14/sistema-bancario/.gitignore
@@ -1,0 +1,2 @@
+node_modules
+ToDo.txt
