+++ semana14/sistema-bancario/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "ts-node-dev ./src/index.ts",
-    "build": "node ./build/index.ts"
+    "build": "tsc && node ./build/index.js"
   },
   "keywords": [],
   "author": "",
